--- a/src/h.js
+++ b/src/h.js
@@ -1,28 +1,18 @@
-function addChild(children, val) {
-  if (val == null || typeof val === 'boolean') return
+var isArray = Array.isArray
 
-  if (Array.isArray(val)) {
-    for (var j = 0; j < val.length; j++) {
-      children.push(val[j])
-    }
-  }
-  else {
-    children.push(typeof val === 'number' ? val + '' : val)
-  }
-}
-
-export function h(tag, data, values) {
+export function createNode(tag, data, values) {
   var children = []
 
-  if (arguments.length > 2) {
-    if (Array.isArray(values)) {
-      for (var i = 0; i < values.length; i++) {
-        addChild(children, values[i])
+  for (var i = 0; i < values.length; i++) {
+    var val = values[i]
+    if (val && val !== true && val !== false) {
+      if (isArray(val)) {
+        for (var j = 0; j < val.length; j++) {
+          children.push(val[j])
+        }
       }
-    }
-    else {
-      for (var i = 2; i < arguments.length; i++) {
-        addChild(children, arguments[i])
+      else {
+        children.push(typeof val === 'number' ? val + '' : val)
       }
     }
   }
@@ -34,4 +24,15 @@ export function h(tag, data, values) {
         children: children
       }
     : tag(data, children)
+}
+
+export function h(tag, data, values) {
+  if (!isArray(values)) {
+    values = []
+    for (var i = 2; i < arguments.length; i++) {
+      values.push(arguments[i])
+    }
+  }
+
+  return createNode(tag, data, values || [])
 }