The child of #187.

**Before**
```sh
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~319 ms (0 s + 319243686 ns)

# vnode with a single child 1M times
ok ~1.48 s (1 s + 481573267 ns)

# vnode with a single array child 1M times
ok ~1.67 s (1 s + 670888960 ns)

all benchmarks completed
ok ~3.47 s (3 s + 471705913 ns)
```

**After**
```sh
NANOBENCH version 2
> node

# vnode with no child 1M times
ok ~330 ms (0 s + 330196882 ns)

# vnode with a single child 1M times
ok ~954 ms (0 s + 953737873 ns)

# vnode with a single array child 1M times
ok ~990 ms (0 s + 990082693 ns)

all benchmarks completed
ok ~2.27 s (2 s + 274017448 ns)
```

I've committed a small facility to compare perfs, I'll strip them before squashing and merging.

Things to explore:
 - [ ] Better bench use cases. @jbucaran ideas?
 - [ ] Node constructors (https://github.com/hyperapp/hyperapp/pull/187#issuecomment-294316864)
 - [ ] Recycling.

Other `h` out there:
 - **preact**: https://github.com/developit/preact/blob/master/src/h.js
 - **react**: https://github.com/facebook/react/blob/5c6a496d98b80d19d851c2ec3a56d760b7118a50/src/isomorphic/classic/element/ReactElement.js#L180
 - **inferno**: https://github.com/infernojs/inferno/blob/master/packages/inferno-create-element/src/index.ts

Refs:
 - Why 2 for loops: https://github.com/hyperapp/hyperapp/pull/187#issuecomment-294339008