@ngryman 2017-07-13
Are we ok with this implementation?
If so, I can add tests to get 100% coverage and merge it.

As discussed with @jbucaran node constructors and recycling will be introduced in another PR.