@jbucaran 2017-07-13
@ngryman Thanks, but did you run these changes through the js-framework-benchmarks? That's the indicator I've used before making "big" changes that could affect performance. I know you are trying to improve `h` here, so I am optimistic.

Let me know the results if and when you have them! 👍