@jbucaran 2017-07-26
@ngryman Did ever have a chance to run these through the benchmarks? 🤔 I'd really like to make a decision here! 👋😄