@ngryman 2017-07-26
@jbucaran Hey I gave you the result on Slack 15 days ago 😋
They were approx 7% better, which is not a game changer, but I didn't expect more from `h`.

<img width="202" alt="screen shot 2017-07-13 at 20 19 41" src="https://user-images.githubusercontent.com/892048/28646105-4a0cdaa4-7260-11e7-83a4-42b38d753ccc.png">