@ngryman 2017-07-26
@jbucaran We could also gain bytes and perf if this babel issue is resolved: https://github.com/babel/babel/issues/2034#issuecomment-294189475

The guy seemed open to a PR for this. I might get some free time next week, I think I'm gonna push this, it would also make lots of library authors happier.