@leeoniya 2017-07-26
> They were approx 7% better, which is not a game changer,

7% is actually quite a lot