@SkaterDad 2017-07-26
@ngryman  That JSX transform fix would be great.  When I first looked at 'h()', that arguments stuff struck me as really strange!