@jbucaran 2017-07-27
@ngryman Those results looked great, so I setup js-frameworks-benchmarks again from scratch and here are my results using the h.js currently on master VS your h.js (I made them both use the same app.js to avoid false positives).

<img width="105" alt="screen shot 2017-07-27 at 21 44 13" src="https://user-images.githubusercontent.com/56996/28670685-c4891568-7314-11e7-94eb-4f4e8fad7efd.png">

But like @SkaterDad is saying, that JSX transform fix would be really great.

Not sure what else to do.