@jbucaran 2017-07-27
I ran it --count 5, which took a while and the results were better:

<img width="107" alt="screen shot 2017-07-27 at 21 56 22" src="https://user-images.githubusercontent.com/56996/28671123-80c3abf2-7316-11e7-8c1b-ebb8864cf147.png">