@ngryman 2017-07-27
@jbucaran I'm trying to understand, you are comparing `0.11.0` with `0.9.1`. Where is the optimized `h`? Those two versions have lots of other additions.