@ngryman 2017-07-27
@jbucaran Surprised, I'll take a look. The fact that the impact is not visible enough in this benchmark wouldn't surprise me, but your first batch of result is very... interesting :D