@ngryman 2017-07-27
I've run several benchmarks, several times and got the same results for both implementations. I still can't explain why `0.11` is the same and `0.9` had a real gap between the two.

I have to admit that starting optimization passes with `h` was not the best idea. I've done it because it was quick and fun. But it does not really bring a lot to the table right now... I kind of expected that, but I was hoping for a slight perf boost like I measured in `0.9`.

I should have started elsewhere: normalizing `vnode`s.

The proposed implementation is faster because potentially optimizable by the JIT. Currently in a real world use case, we don't leverage that optimization because our `vnode`s are of several shapes, ... and we have bigger problems.
When we will have a "typed" `vtree`, we'll probably have a little performance boost from this PR, mainly in animations where `h` could potentially become hot. But again, this will be tiny.

I would propose to focus on normalizing our `vnode`s and recycle them (because we are generating tons of garbage right now) and revisit this PR later when those optimizations are done.